+++ test/unit/textbox.js
@@ -55,6 +55,9 @@
     strokeUniform: false,
     path: null,
     direction: 'ltr',
+    pathStartOffset: 0,
+    pathSide: 'left',
+    pathAlign: 'baseline'
   };
 
   QUnit.test('constructor', function(assert) {